Move relative code to `pprof.go`
